Duration bag was fixed

+++ src/main/java/com/training/audiomanager/command/track/AddTrack.java
@@ -35,7 +35,7 @@ public class AddTrack implements Command {
         Long genreId = inputUtil.inputLongValue(request, ParameterConstants.GENRE_ID);
         String album = inputUtil.inputStringValue(request, ParameterConstants.ALBUM);
         String name = inputUtil.inputStringValue(request, ParameterConstants.NAME);
-        double duration = inputUtil.inputDoubleValue(request, ParameterConstants.DURATION);
+        int duration = inputUtil.inputDoubleValue(request, ParameterConstants.DURATION);
         if (!inputUtil.isValid()) {
             return PageConstants.ADD_TRACK_PAGE_REDIRECT;
         }
